requirements.txt
python-dotenv==1.0.0
langchain==0.0.137 
# langchain[all]==0.0.137 # Use [all] to get full installation, approx ~6GB
# pinecone-client==2.2.1
openai==0.27.8
tiktoken==0.4.0 